fix: move drag to wrapper so PR card follows FAB position
Move drag responsibility from FabButton to the parent motion.div using
useDragControls, so the card and button move together when dragged.
Also replace useState with useMemo in useFabPosition to avoid
unnecessary re-renders since the state was only read once to seed
motion values.
