+++ src/modules/gym-fab/components/fab-button/index.tsx
@@ -1,50 +1,27 @@
 'use client';
 
 import { Dumbbell } from 'lucide-react';
-import { motion, type MotionValue } from 'motion/react';
-import { type RefObject, useRef } from 'react';
+import { motion } from 'motion/react';
+import { type PointerEvent } from 'react';
 
 const springTransition = { type: 'spring', bounce: 0.2, duration: 0.4 };
 
 interface FabButtonProps {
   isOpen: boolean;
   onToggle: () => void;
-  x: MotionValue<number>;
-  y: MotionValue<number>;
-  dragConstraints: RefObject<HTMLDivElement | null>;
-  onDragEnd: () => void;
+  onPointerDown: (e: PointerEvent) => void;
 }
 
 export const FabButton = ({
   isOpen,
   onToggle,
-  x,
-  y,
-  dragConstraints,
-  onDragEnd,
+  onPointerDown,
 }: FabButtonProps) => {
-  const isDragging = useRef(false);
-
   return (
     <motion.button
       className="bg-primary text-primary-foreground flex size-12 cursor-grab items-center justify-center rounded-full shadow-lg active:cursor-grabbing"
-      style={{ x, y }}
-      drag
-      dragMomentum={false}
-      dragElastic={0.1}
-      dragConstraints={dragConstraints}
-      onDragStart={() => {
-        isDragging.current = true;
-      }}
-      onDragEnd={() => {
-        onDragEnd();
-        requestAnimationFrame(() => {
-          isDragging.current = false;
-        });
-      }}
-      onClick={() => {
-        if (!isDragging.current) onToggle();
-      }}
+      onPointerDown={onPointerDown}
+      onClick={onToggle}
       whileHover={{ scale: 1.1 }}
       whileTap={{ scale: 0.95 }}
       aria-label={isOpen ? 'Close gym PRs' : 'Show gym PRs'}
